+++ src/chains/foundry.ts
@@ -72,6 +72,8 @@ export const foundryChainNameMap: Record<ChainId, string> = {
   [ChainId.FLW]: 'flow',
   [ChainId.HMI]: 'hemi',
   [ChainId.MON]: 'monad',
+  [ChainId.STA]: 'stable',
+  [ChainId.MEG]: 'megaeth',
 
   // Custom chains
   [7777777 as ChainId]: 'zora',